Restructure api-key command
