+++ backend/cli.py
@@ -21,11 +21,10 @@ os.environ.setdefault("DJANGO_SETTINGS_MODULE", "backend.settings")  # Adjust as
 django.setup()
 
 # Importing Models, adjust as needed
-from restapi.models import Mission, Tag, Mission_tags  # noqa
-from restapi.serializer import MissionSerializer, TagSerializer  # noqa
-from rest_framework_api_key.models import APIKey  # noqa
+from restapi.models import Mission  # noqa
 from cli_commands.MissionCommand import MissionCommand  # noqa
 from cli_commands.TagCommand import TagCommand  # noqa
+from cli_commands.ApiKeyCommand import ApiKeyCommand  # noqa
 
 # Set up logging
 logging.basicConfig(level=logging.INFO)
@@ -125,64 +124,6 @@ def sync_folder(folder_path, location=None, notes=None):
         logging.error("invalid path")
 
 
-def add_api_key(name, expiry_date=None):
-    """
-    Add/create a new API KEY with an optional expiration date.
-    The date is expected to be a date and time, but can also be just a date.
-
-    Args:
-        name: Name of new API
-        expiry_date (optional): Expiration date. Defaults to None.
-    """
-    try:
-        name, key = APIKey.objects.create_key(name=name, expiry_date=expiry_date)
-    except Exception as e:
-        logging.error(f"Couldn't create API KEY with name '{name}': {e}")
-        return
-
-    logging.info(
-        f"Api Key '{name}' created\nkey: '{key}'\nThis key will not be visible again!"
-    )
-
-
-def remove_api_key(prefix=None, name=None):
-    """"
-    Remove an API KEY either by name or prefix.\\
-    The prefix can be found with `api-key list`.\\
-    Both are optional but one has to be given.\\
-    If both are given the prefix is preferred, since it's unique.\\
-    If the name is used and there are mutliple entries with the same name, all are removed.
-    
-    Args:
-        prefix (optional): the prefix of an API KEY
-        name (optional): the name of one or more API KEYs
-    """
-    if prefix:
-        api_key = APIKey.objects.filter(prefix=prefix)
-    elif name:
-        api_key = APIKey.objects.filter(name=name)
-    else:
-        logging.error("Either 'prefix' or 'name' must be provided")
-        return
-
-    if not api_key.exists():
-        logging.error(f"API KEY '{prefix if prefix else name}' not found")
-        return
-
-    try:
-        api_key.delete()
-    except Exception as e:
-        logging.error(f"Couldn't remove API KEY: {e}")
-        return
-
-    logging.info(f"Removed API KEY '{prefix if prefix else name}'")
-
-
-def list_api_keys():
-    keys = APIKey.objects.values()
-    Command.print_table(keys)
-
-
 class Interactive(code.InteractiveConsole):
     def __init__(self, help):
         super().__init__(locals=None, filename="<console>")
@@ -242,18 +183,6 @@ def interactive(parser: argparse.ArgumentParser):
             )
 
 
-def api_key_command(api_key_parser, args):
-    match args.api_key:
-        case "add":
-            add_api_key(args.name, args.expiry_date)
-        case "remove":
-            remove_api_key(args.prefix, args.name)
-        case "list":
-            list_api_keys()
-        case _:
-            api_key_parser.print_help()
-
-
 def folder_arg_parser(subparser):
     """
     Parser setup for addfolder subcommand
@@ -275,31 +204,10 @@ def sync_arg_parser(subparser):
     sync_parser.add_argument("--notes", required=False, help="other mission details")
 
 
-def api_key_arg_parser(subparser: argparse._SubParsersAction):
-    api_key_parser = subparser.add_parser("api-key", help="Modify API-KEYS")
-    api_key_subparser = api_key_parser.add_subparsers(dest="api_key")
-
-    # Add command
-    add_parser = api_key_subparser.add_parser("add", help="Add API KEY")
-    add_parser.add_argument("--name", required=True, help="Name of API KEY")
-    add_parser.add_argument(
-        "--expiry-date", required=False, help="Set expiry date for key"
-    )
-
-    # Remove command
-    remove_parser = api_key_subparser.add_parser("remove", help="Remove API KEY")
-    remove_parser.add_argument("--name", required=False, help="Name of API KEY")
-    remove_parser.add_argument("--prefix", required=False, help="Prefix of API KEY")
-
-    # List command
-    _ = api_key_subparser.add_parser("list", help="List all API KEYS")
-
-    return api_key_parser
-
-
 commands: dict[str, Command] = {
     "mission": MissionCommand(),
     "tag": TagCommand(),
+    "api-key": ApiKeyCommand(),
 }
 
 
@@ -315,8 +223,6 @@ def main(args):
 
     sync_arg_parser(subparser)
 
-    api_key_parser = api_key_arg_parser(subparser)
-
     argcomplete.autocomplete(parser)
 
     args = parser.parse_args(args)
@@ -327,8 +233,6 @@ def main(args):
             add_mission_from_folder(args.path, args.location, args.notes)
         case "syncfolder":
             sync_folder(args.path, args.location, args.notes)
-        case "api-key":
-            api_key_command(api_key_parser, args)
         case _:
             if args.command in commands:
                 commands[args.command].command(args)
